moved modal state to OrderModal moved order button to OrderModal moved order button styling to ordermodal.css
