+++ src/components/controls/control.css
@@ -6,10 +6,7 @@
   width: 100%;
 }
 
-.btn-order{
-  background-color: #DAD735;
-  width: 40%;
-}
+
 
 .outline{
   border: 3px dotted blue;
