+++ src/components/controls/orderModal/OrderModal.jsx
@@ -1,29 +1,44 @@
+import { useState } from "react";
 import Modal from "react-bootstrap/Modal";
 
 import { foodItemsRate } from "../../helpers/helper";
 
 import "./orderModal.css";
 
-export default function OrderModal({ show, foodPrice, price, hide }) {
+export default function OrderModal({ foodPrice, price }) {
+  const [show, setShow] = useState(false);
+
+  const showModal = () => setShow(true);
+  const hideModal = () => setShow(false);
+
   const quantities = Object.entries(foodPrice).map(([key, val]) => (
     <li>
       {key}: $ {val / foodItemsRate[key]}
     </li>
   ));
   return (
-    <Modal show={show} onHide={hide}>
-      <Modal.Body>
-        <h3>Your Order Summary</h3>
-        <ul>{quantities}</ul>
-        <strong>Total Price: ${price}</strong>
-        <p>Continue to Checkout?</p>
-        <button className="btn-cancel" onClick={hide}>
-          Cancel
-        </button>
-        <button className="btn-continue" onClick={hide}>
-          Continue
-        </button>
-      </Modal.Body>
-    </Modal>
+    <>
+      <button
+        className="justify-content-center btn-order"
+        onClick={showModal}
+        disabled={price === 3}
+      >
+        ORDER
+      </button>
+      <Modal show={show} onHide={hideModal}>
+        <Modal.Body>
+          <h3>Your Order Summary</h3>
+          <ul>{quantities}</ul>
+          <strong>Total Price: ${price.toFixed(2)}</strong>
+          <p>Continue to Checkout?</p>
+          <button className="btn-cancel" onClick={hideModal}>
+            Cancel
+          </button>
+          <button className="btn-continue" onClick={hideModal}>
+            Continue
+          </button>
+        </Modal.Body>
+      </Modal>
+    </>
   );
 }
