+++ src/components/controls/orderModal/orderModal.css
@@ -1,3 +1,9 @@
+
+.btn-order{
+  background-color: #DAD735;
+  width: 40%;
+}
+
 .btn-cancel{
   color: #944317;
 }
